Move hooks outside of the slice for connected sites
