+++ src/components/content-tab-import-export.tsx
@@ -19,7 +19,7 @@ import { cx } from 'src/lib/cx';
 import { getIpcApi } from 'src/lib/get-ipc-api';
 import { getLocalizedLink } from 'src/lib/get-localized-link';
 import { useI18nLocale } from 'src/stores';
-import { useConnectedSitesData } from 'src/stores/sync/connected-sites-slice';
+import { useConnectedSitesData } from 'src/stores/sync';
 
 interface ContentTabImportExportProps {
 	selectedSite: SiteDetails;
